Reorganize projects so common code can sit in the StreamServer project (in CertHelper and Utilities files). Certificate and Utility functions moved out of SSLHelper which completely eliminated SSLHelper for the client and much reduced it for the server. Moved CreateCertificate code into CertHelper and eliminated it. The one surprise was I had to modify the stdafx.h file for SteamServer because it used used a different definitions for CString.
