+++ StreamClient/StreamClient.vcxproj
@@ -93,22 +93,23 @@
     <ClInclude Include="ISocketStream.h" />
     <ClInclude Include="..\StreamServer\SecurityHandle.h" />
     <ClInclude Include="SSLClient.h" />
-    <ClInclude Include="SSLHelper.h" />
     <ClInclude Include="stdafx.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\StreamServer\CertRAII.cpp" />
-    <ClCompile Include="..\StreamServer\CreateCertificate.cpp" />
-    <ClCompile Include="..\StreamServer\Utilities.cpp" />
-    <ClCompile Include="ActiveSock.cpp">
+    <ClCompile Include="..\StreamServer\CertHelper.cpp" />
+    <ClCompile Include="..\StreamServer\CertRAII.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">NotUsing</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="SSLClient.cpp">
+    <ClCompile Include="..\StreamServer\Utilities.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Use</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Use</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="SSLHelper.cpp">
+    <ClCompile Include="ActiveSock.cpp">
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Use</PrecompiledHeader>
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Use</PrecompiledHeader>
+    </ClCompile>
+    <ClCompile Include="SSLClient.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Use</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Use</PrecompiledHeader>
     </ClCompile>
