+++ StreamClient/StreamClient.vcxproj.filters
@@ -26,9 +26,6 @@
     <ClInclude Include="EventWrapper.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="SSLHelper.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\StreamServer\SecurityHandle.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -49,16 +46,13 @@
     <ClCompile Include="SSLClient.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="SSLHelper.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="..\StreamServer\Utilities.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\StreamServer\CreateCertificate.cpp">
+    <ClCompile Include="..\StreamServer\CertRAII.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\StreamServer\CertRAII.cpp">
+    <ClCompile Include="..\StreamServer\CertHelper.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
